Add a new.js file and import to index.js
