+++ src/index.html
@@ -13,7 +13,7 @@
         </header>
         <main>
            <div class="input-container"><input type="text" placeholder="Add your list..." class="task-input" ><span class="enter-btn">&#8629;</span></div>
-           <ul class="todays-list" id="todays-list" [data-tasks]></ul>
+           <ul class="todays-list" id="todays-list"></ul>
            <div class="clear-completed"><a id="task-completed" href="#">Clear Completed</a></div>
         </main>
     </div>
